Revert "Don 704 move secondary column into separate component"
