--- src/app/donation-start/donation-start-container/donation-start-container.component.html
@@ -1,20 +1,0 @@
-<main class="b-container">
-  <div class="b-back-button">
-    <a mat-icon-button [routerLink]="'/campaign/' + campaign.id">
-      <mat-icon aria-hidden="false" aria-label="Back">keyboard_arrow_left</mat-icon>
-      Back to campaign
-    </a>
-  </div>
-  <div class="c-form-container">
-    <div class="b-primary-column">
-      <app-donation-start-primary
-      [campaign]="campaign"
-    />
-    </div>
-    <div class="b-secondary-column">
-      <app-donation-start-secondary
-      [campaign]="campaign"
-    />
-    </div>
-  </div>
-</main>
