--- src/app/donation-start/donation-start-secondary/donation-start-secondary.component.html
@@ -1,68 +1,0 @@
-<div *ngIf="campaign">
-    <div id="campaign-highlights-beneficiaries-and-categories">
-    <!-- Content of this DIV is copied from src/app/campaign-info/campaign-info.component.html . I'm not happy about
-            copying and pasting, but when I (bdsl) tried using that component from here I couldn't get the styling to look
-            anything like how it does on the donation info page. This way I can get it identical, other than very minor varitions in
-            sizes of icons
-            -->
-    <div>
-
-        <div *ngIf="campaign.isMatched" class="match-funded-badge-wrapper">
-        <div class="match-funded-badge">
-            Match funded
-        </div>
-        </div>
-        <div class="c-banner">
-        <img
-            class="c-banner__image b-w-100"
-            *ngIf="bannerUri"
-            [src]="bannerUri"
-            alt=""
-            role="presentation"
-        >
-        </div>
-
-        <biggive-campaign-highlights
-        space-below=0
-        primary-figure-label="Amount raised"
-        [primaryFigureAmount]="campaignRaised"
-        secondary-figure-label="Campaign target"
-        [secondaryFigureAmount]="campaignTarget"
-        [progressBarCounter]="getPercentageRaised(campaign)?.toFixed(0)"
-        primary-stat-icon="AlarmClock"
-        [primaryStatText]="campaignFinished ? 'Closed ' + datePipe.transform(campaign.endDate, 'd LLL yyyy') : ( campaignOpen ? (timeLeftPipe.transform(campaign.endDate) + ' left') : 'Opens in ' + timeLeftPipe.transform(campaign.startDate))"
-        secondary-stat-icon="Heart"
-        [secondaryStatText]="campaign.donationCount === 1 ? '1 donation' : campaign.donationCount + ' donations'"
-        [championName]="campaign.championName"
-        [championUrl]="'/' + campaign.parentRef + '/' + campaign.championRef">
-        </biggive-campaign-highlights>
-    </div>
-    </div>
-</div>
-<div style="margin-bottom: 30px;">
-<h5>Categories</h5>
-<hr>
-<div *ngIf="campaign">
-    <div class="icons">
-        <li *ngFor="let category of campaign.categories" class="icon-wrapper">
-        <fa-icon [title]="category" [icon]="getCategoryIcon(category)"></fa-icon>
-        <span class="c-icon-title">
-            <!-- add zero width space to allow line break -->
-            {{ category.split('/').join('/\u200B') }}
-        </span>
-        </li>
-    </div>
-
-    <h5>Beneficiaries</h5>
-    <hr>
-    <div class="icons">
-        <li *ngFor="let beneficiary of campaign.beneficiaries" class="icon-wrapper">
-        <fa-icon [title]="beneficiary" [icon]="getBeneficiaryIcon(beneficiary)"></fa-icon>
-        <span class="c-icon-title">
-            <!-- add zero width space to allow line break -->
-            {{ beneficiary.split('/').join('/\u200B') }}
-        </span>
-        </li>
-    </div>
-</div>
-</div>
