--- src/app/donation-start/donation-start-secondary/donation-start-secondary.component.scss
@@ -1,25 +1,0 @@
-@import '../../../abstract';
-
-h5 {
-    font-size: 1rem;
-}
-
-.icons {
-    @include format-icon-container;
-}
-
-.icon-wrapper {
-    @include format-icon-list-item;
-
-    // in campaign-details.component.scss the selector here is `svg` rather than `fa-icon`, but that doesn't seem to work
-    // here.
-    fa-icon.ng-fa-icon {
-        color: $colour-primary; // #2C089BFF
-        font-size: 40px;
-        min-width: 4rem;
-    }
-
-    .c-icon-title {
-        @include format-icon-title;
-    }
-}
