--- src/app/donation-start/donation-start-secondary/donation-start-secondary.component.spec.ts
@@ -1,23 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { DonationStartSecondaryComponent } from './donation-start-secondary.component';
-
-describe('DonationStartSecondaryComponent', () => {
-  let component: DonationStartSecondaryComponent;
-  let fixture: ComponentFixture<DonationStartSecondaryComponent>;
-
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      declarations: [ DonationStartSecondaryComponent ]
-    })
-    .compileComponents();
-
-    fixture = TestBed.createComponent(DonationStartSecondaryComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
