+++ src/app/donation-start/donation-start.component.ts
@@ -25,47 +25,48 @@ import { debounceTime, distinctUntilChanged, retryWhen, startWith, switchMap, ta
 import { PaymentMethod, StripeCardElement, StripeElementChangeEvent, StripeError, StripePaymentRequestButtonElement } from '@stripe/stripe-js';
 import { EMPTY, Observer } from 'rxjs';
 
-import { AnalyticsService } from '../../analytics.service';
-import { Campaign } from '../../campaign.model';
-import { CampaignService } from '../../campaign.service';
-import { CardIconsService } from '../../card-icons.service';
-import { COUNTRIES } from '../../countries';
-import { Donation } from '../../donation.model';
-import { DonationCreatedResponse } from '../../donation-created-response.model';
-import { DonationService } from '../../donation.service';
-import { DonationStartMatchConfirmDialogComponent } from '../donation-start-match-confirm-dialog.component';
-import { DonationStartMatchingExpiredDialogComponent } from '../donation-start-matching-expired-dialog.component';
-import { DonationStartOfferReuseDialogComponent } from '../donation-start-offer-reuse-dialog.component';
-import { environment } from '../../../environments/environment';
-import { ExactCurrencyPipe } from '../../exact-currency.pipe';
-import { GiftAidAddress } from '../../gift-aid-address.model';
-import { GiftAidAddressSuggestion } from '../../gift-aid-address-suggestion.model';
-import { IdentityService } from '../../identity.service';
-import { LoginModalComponent } from '../../login-modal/login-modal.component';
-import { MetaPixelService } from '../../meta-pixel.service';
-import { PageMetaService } from '../../page-meta.service';
-import { Person } from '../../person.model';
-import { PostcodeService } from '../../postcode.service';
-import { retryStrategy } from '../../observable-retry';
-import { StripeService } from '../../stripe.service';
-import { getCurrencyMaxValidator } from '../../validators/currency-max';
-import { getCurrencyMinValidator } from '../../validators/currency-min';
-import { EMAIL_REGEXP } from '../../validators/patterns';
-import { ValidateBillingPostCode } from '../../validators/validate-billing-post-code';
-import {CampaignGroupsService} from "../../campaign-groups.service";
-import {TimeLeftPipe} from "../../time-left.pipe";
+import { AnalyticsService } from '../analytics.service';
+import { Campaign } from '../campaign.model';
+import { CampaignService } from '../campaign.service';
+import { CardIconsService } from '../card-icons.service';
+import { COUNTRIES } from '../countries';
+import { Donation } from '../donation.model';
+import { DonationCreatedResponse } from '../donation-created-response.model';
+import { DonationService } from '../donation.service';
+import { DonationStartMatchConfirmDialogComponent } from './donation-start-match-confirm-dialog.component';
+import { DonationStartMatchingExpiredDialogComponent } from './donation-start-matching-expired-dialog.component';
+import { DonationStartOfferReuseDialogComponent } from './donation-start-offer-reuse-dialog.component';
+import { environment } from '../../environments/environment';
+import { ExactCurrencyPipe } from '../exact-currency.pipe';
+import { GiftAidAddress } from '../gift-aid-address.model';
+import { GiftAidAddressSuggestion } from '../gift-aid-address-suggestion.model';
+import { IdentityService } from '../identity.service';
+import { LoginModalComponent } from '../login-modal/login-modal.component';
+import { MetaPixelService } from '../meta-pixel.service';
+import { PageMetaService } from '../page-meta.service';
+import { Person } from '../person.model';
+import { PostcodeService } from '../postcode.service';
+import { retryStrategy } from '../observable-retry';
+import { StripeService } from '../stripe.service';
+import { getCurrencyMaxValidator } from '../validators/currency-max';
+import { getCurrencyMinValidator } from '../validators/currency-min';
+import { EMAIL_REGEXP } from '../validators/patterns';
+import { ValidateBillingPostCode } from '../validators/validate-billing-post-code';
+import {CampaignGroupsService} from "../campaign-groups.service";
+import {TimeLeftPipe} from "../time-left.pipe";
+import {ImageService} from "../image.service";
 import { MatomoTracker } from 'ngx-matomo';
 
 @Component({
-  selector: 'app-donation-start-primary',
-  templateUrl: './donation-start-primary.component.html',
-  styleUrls: ['./donation-start-primary.component.scss'],
+  selector: 'app-donation-start',
+  templateUrl: './donation-start.component.html',
+  styleUrls: ['./donation-start.component.scss'],
   providers: [
     CurrencyPipe,
     TimeLeftPipe,
   ]
 })
-export class DonationStartPrimaryComponent implements AfterContentChecked, AfterContentInit, OnDestroy, OnInit {
+export class DonationStartComponent implements AfterContentChecked, AfterContentInit, OnDestroy, OnInit {
   @ViewChild('captcha') captcha: RecaptchaComponent;
   @ViewChild('idCaptcha') idCaptcha: RecaptchaComponent;
   @ViewChild('cardInfo') cardInfo: ElementRef;
@@ -79,9 +80,7 @@ export class DonationStartPrimaryComponent implements AfterContentChecked, After
   requestButtonShown = false;
   showChampionOptIn = false;
 
-  @Input() campaign: Campaign;
-  @Input() column: 'primary'|'secondary'
-
+  campaign: Campaign;
   donation?: Donation;
 
   campaignOpenOnLoad: boolean;
@@ -132,6 +131,9 @@ export class DonationStartPrimaryComponent implements AfterContentChecked, After
   // Track 'Next' clicks so we know when to show missing radio button error messages.
   triedToLeaveGiftAid = false;
   triedToLeaveMarketing = false;
+  campaignFinished: boolean;
+  campaignOpen: boolean;
+  bannerUri: string | null;
   showAllPaymentMethods: boolean = false;
 
   private campaignId: string;
@@ -164,6 +166,8 @@ export class DonationStartPrimaryComponent implements AfterContentChecked, After
 
   private idCaptchaCode?: string;
   private stripeResponseErrorCode?: string; // stores error codes returned by Stripe after callout
+  campaignRaised: string; // Formatted
+  campaignTarget: string; // Formatted
   private stepChangedBlockedByCaptcha = false;
 
 
@@ -176,6 +180,7 @@ export class DonationStartPrimaryComponent implements AfterContentChecked, After
     @Inject(ElementRef) private elRef: ElementRef,
     private formBuilder: FormBuilder,
     private identityService: IdentityService,
+    private imageService: ImageService,
     private matomoTracker: MatomoTracker,
     private metaPixelService: MetaPixelService,
     private pageMeta: PageMetaService,
@@ -208,6 +213,7 @@ export class DonationStartPrimaryComponent implements AfterContentChecked, After
       this.stripeService.init();
     }
 
+    this.campaign = this.route.snapshot.data.campaign;
     this.setCampaignBasedVars();
 
     const idAndJWT = this.identityService.getIdAndJWT();
@@ -216,6 +222,15 @@ export class DonationStartPrimaryComponent implements AfterContentChecked, After
         this.loadAuthedPersonInfo(idAndJWT.id, idAndJWT.jwt);
       }
     }
+
+    this.imageService.getImageUri(this.campaign.bannerUri, 830).subscribe(uri => this.bannerUri = uri);
+
+    // This block of code is copied from campaign-info.component. Apologies for duplication.
+    this.campaignTarget = this.currencyPipe.transform(this.campaign.target, this.campaign.currencyCode, 'symbol', '1.0-0') as string;
+    this.campaignRaised = this.currencyPipe.transform(this.campaign.amountRaised, this.campaign.currencyCode, 'symbol', '1.0-0') as string;
+    this.campaignFinished = CampaignService.isInPast(this.campaign);
+    this.campaignOpen = CampaignService.isOpenForDonations(this.campaign);
+
 
     const formGroups: {
       amounts: FormGroup,   // Matching reservation happens at the end of this group.
@@ -1929,6 +1944,15 @@ export class DonationStartPrimaryComponent implements AfterContentChecked, After
       donation.donationAmount + donation.tipAmount + donation.feeCoverAmount,
       donation.donationAmount,
     );
+  }
+
+  // Three functions below copied from campaign-info.component. Apologies for duplication.
+  getBeneficiaryIcon(beneficiary: string) {
+    return CampaignGroupsService.getBeneficiaryIcon(beneficiary);
+  }
+
+  getCategoryIcon(category: string) {
+    return CampaignGroupsService.getCategoryIcon(category);
   }
 
   getPercentageRaised(campaign: Campaign): number | undefined {
